# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Sistema de Controle Antimicrobianos Veterinários",
    "description": "Arquitetura de um Sistema de Controle de Uso, Vendas e Prescrição de Antimicrobianos Veterinários"
  },
  "security": [{"bearer":[]}],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/ceCategory": {
      "get": {
        "summary": "Show all commercial establishments category",
        "tags": [
          "Commercial Establishments Category"
        ],
        "responses": {
          "200": {
            "description": "Return all commercial establishments category"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an new commercial establishments category",
        "tags": [
          "Commercial Establishments Category"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/CeCategory/Create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Commercial establishments category created"
          },
          "400": {
            "description": "Error when creating commercial establishments category"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update an commercial establishments category",
        "tags": [
          "Commercial Establishments Category"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/CeCategory/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Commercial establishments category Deleted"
          },
          "400": {
            "description": "Error when deleting commercial establishments category"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "delete": {
        "summary": "Delete a commercial establishments category",
        "tags": [
          "Commercial Establishments Category"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/CeCategory/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Commercial establishments category updated"
          },
          "400": {
            "description": "Error when delete commercial establishments category"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/oieClassification": {
      "get": {
        "summary": "Show all oie classification",
        "tags": [
          "Oie Classification"
        ],
        "responses": {
          "200": {
            "description": "Return all oie classification"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an new oie classification",
        "tags": [
          "Oie Classification"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/OieClassification/Create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oie classification created"
          },
          "400": {
            "description": "Error when creating oie classification"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update an oie classification",
        "tags": [
          "Oie Classification"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/OieClassification/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oie classification Deleted"
          },
          "400": {
            "description": "Error when deleting oie classification"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "delete": {
        "summary": "Delete a oie classification",
        "tags": [
          "Oie Classification"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/OieClassification/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oie classification updated"
          },
          "400": {
            "description": "Error when delete oie classification"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/omsClassification": {
      "get": {
        "summary": "Show all Oms classification",
        "tags": [
          "Oms Classification"
        ],
        "responses": {
          "200": {
            "description": "Return all Oms classification"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an new Oms classification",
        "tags": [
          "Oms Classification"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/OmsClassification/Create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oms classification created"
          },
          "400": {
            "description": "Error when creating oms classification"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update an Oms classification",
        "tags": [
          "Oms Classification"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/OmsClassification/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oms classification Deleted"
          },
          "400": {
            "description": "Error when deleting oms classification"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "delete": {
        "summary": "Delete a oms classification",
        "tags": [
          "Oms Classification"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/OmsClassification/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oms classification updated"
          },
          "400": {
            "description": "Error when delete oms classification"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/assignedCe": {
      "get": {
        "summary": "Show all assigned Ce",
        "tags": [
          "Assigned Ce"
        ],
        "responses": {
          "200": {
            "description": "Return all assigned Ce"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an new assigned Ce",
        "tags": [
          "Assigned Ce"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/AssignedCe/Create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assigned Ce created"
          },
          "400": {
            "description": "Error when creating assigned Ce"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "put": {
        "summary": "Update an assigned Ce",
        "tags": [
          "Assigned Ce"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/AssignedCe/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assigned Ce Deleted"
          },
          "400": {
            "description": "Error when deleting assigned Ce"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "delete": {
        "summary": "Delete a assigned Ce",
        "tags": [
          "Assigned Ce"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/AssignedCe/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assigned Ce updated"
          },
          "400": {
            "description": "Error when delete assigned Ce"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}